feat(demo): 示例页改为一份真实加密分享 + 口令随 URL fragment 送达
上一版给查看器加了 demo 专用的数据节点读取分支,结果 CodeQL 报出两条 high
(js/xss-through-dom):DOM 文本经 JSON.parse 直达 innerHTML,是一条完整污点路径
——真实分享那条被 Web Crypto 解密打断,所以从没报过。这两条确由本改动引入。

改成示例页**就是一份真实加密分享**:同一份查看器、同一套 AES-256-GCM、同一个
口令流程,只有数据换成公开演示数据集。于是:
- 查看器里**不再有任何 demo 专用代码**(demo-data 读取与注入占位全部删除),
  污点路径消失,告警随之消失;
- 「示例与生产一致」不再是机制相似,而是同一条代码路径的产物。

顺带落地一项生产能力:**口令可随 URL fragment 送达**(`#<口令>`)。`#` 之后按
HTTP 规范不上行,密钥只存在于收件人浏览器本地 —— 这正是后续二维码/链接分享
(#58)所需。到达即自动解密,并清除 hash,避免口令留在地址栏与浏览历史。
实测:解析期与 load 时清除都会被浏览器随后的片段导航覆盖,故补一次延迟清除。

「示例数据」横幅改为生成时注入的**静态 HTML**,不碰查看器脚本 —— 改脚本会使
CSP 锁死的 sha256 失配、整段脚本被拒绝执行。
