+++ web/hosted-viewer/index.html
@@ -12,7 +12,7 @@
         a=d.find(o,i)
         if a<0:break
         a+=len(o);b=d.find(c,a);print(base64.b64encode(hashlib.sha256(d[a:b].encode()).digest()).decode());i=b" -->
-<meta http-equiv="Content-Security-Policy" content="default-src 'none'; script-src 'sha256-7Z0+4VB3CnpMGP9NdEXdG4wifm/kpyirE2yqr+CEnyc=' 'sha256-AUW06Y41O6ldTTMMCF/iDphgxQIz4HMqvI0BzthK0NU='; style-src 'unsafe-inline'; img-src data:; font-src data:; frame-src data:; connect-src 'none'; form-action 'none'; base-uri 'none'; frame-ancestors 'none'">
+<meta http-equiv="Content-Security-Policy" content="default-src 'none'; script-src 'sha256-7Z0+4VB3CnpMGP9NdEXdG4wifm/kpyirE2yqr+CEnyc=' 'sha256-J1H0VWFrSXHrv/qNRbQwBdey5DPvFugU/irSMShgWBM='; style-src 'unsafe-inline'; img-src data:; font-src data:; frame-src data:; connect-src 'none'; form-action 'none'; base-uri 'none'; frame-ancestors 'none'">
 <meta name="viewport" content="width=device-width, initial-scale=1">
 <title>MedMe 加密分享查看器</title>
 <link rel="icon" href="data:image/svg+xml,%3Csvg%20xmlns%3D%22http%3A%2F%2Fwww.w3.org%2F2000%2Fsvg%22%20viewBox%3D%220%200%201024%201024%22%3E%0A%3Cdefs%3E%3ClinearGradient%20id%3D%22g%22%20x1%3D%220%22%20y1%3D%220%22%20x2%3D%221%22%20y2%3D%221%22%3E%3Cstop%20offset%3D%220%22%20stop-color%3D%22%232bc0d2%22%2F%3E%3Cstop%20offset%3D%220.52%22%20stop-color%3D%22%231789c1%22%2F%3E%3Cstop%20offset%3D%221%22%20stop-color%3D%22%231560a8%22%2F%3E%3C%2FlinearGradient%3E%3C%2Fdefs%3E%0A%3Crect%20width%3D%221024%22%20height%3D%221024%22%20rx%3D%22232%22%20fill%3D%22url%28%23g%29%22%2F%3E%0A%3Cpath%20transform%3D%22translate%2885.76%2C804.70%29%20scale%280.82760%2C-0.82760%29%22%20d%3D%22M326%20640C320%20672%20323%20688%20335%20688C431%20690%20535%20708%20650%20741C668%20748%20687%20734%20704%20699C720%20664%20716%20650%20691%20658C672%20666%20641%20663%20600%20651C557%20639%20509%20631%20458%20627C405%20623%20371%20615%20355%20605C342%20596%20332%20606%20326%20640ZM212%20393C226%20508%20234%20570%20236%20582C234%20600%20241%20605%20259%20600C274%20594%20292%20586%20310%20578C325%20568%20326%20555%20314%20537C283%20459%20266%20368%20260%20267C252%20165%20254%20101%20266%2075C271%2062%20321%2063%20413%2079C503%2095%20582%20104%20650%20106C717%20108%20757%20111%20772%20115C777%20117%20786%20110%20796%2095C804%2078%20813%2065%20822%2057C830%2047%20832%2034%20829%2020C825%204%20814%200%20798%2010C698%2035%20597%2043%20494%2033C391%2023%20324%2013%20296%203C267%20-6%20246%20-9%20234%20-4C229%20-1%20226%203%20226%2011C226%2018%20220%2040%20210%2077C196%20172%20196%20277%20212%20393ZM567%20292C558%20302%20561%20306%20578%20306C639%20287%20691%20265%20736%20237C740%20234%20741%20219%20737%20189C731%20160%20724%20144%20716%20140C706%20136%20699%20138%20695%20146C689%20154%20683%20165%20675%20179C628%20228%20593%20265%20567%20292ZM761%20328C702%20338%20636%20336%20561%20323C555%20322%20550%20321%20546%20321C538%20294%20525%20267%20509%20240C497%20216%20483%20198%20468%20188C450%20176%20444%20165%20449%20157C451%20149%20434%20142%20397%20135C358%20126%20331%20126%20315%20134C298%20142%20300%20148%20322%20154C324%20154%20335%20159%20355%20169C372%20178%20393%20190%20414%20208C434%20225%20453%20249%20471%20281C473%20291%20476%20300%20480%20309C429%20297%20387%20284%20355%20271C348%20266%20335%20274%20319%20298C302%20321%20310%20330%20342%20324C385%20338%20436%20352%20494%20367C498%20395%20499%20419%20497%20440C489%20438%20480%20435%20471%20431C457%20424%20450%20429%20449%20444C447%20454%20444%20463%20441%20473C422%20448%20400%20427%20377%20408C357%20396%20344%20391%20336%20393C326%20395%20329%20403%20346%20419C385%20473%20409%20513%20421%20540C433%20566%20438%20584%20439%20594C441%20611%20452%20613%20472%20601C489%20587%20502%20577%20511%20571C519%20565%20517%20559%20505%20552C497%20548%20490%20541%20484%20531C480%20525%20475%20517%20467%20507C504%20511%20537%20518%20566%20527C596%20537%20620%20543%20640%20546C649%20547%20656%20538%20660%20520C662%20500%20658%20488%20648%20485C619%20473%20592%20461%20568%20450C564%20449%20563%20448%20562%20447C558%20430%20555%20407%20554%20382C554%20382%20556%20382%20559%20383C598%20391%20635%20396%20671%20402C706%20406%20728%20409%20734%20413C740%20416%20745%20413%20750%20403C754%20394%20760%20387%20768%20383C773%20379%20777%20367%20778%20348C778%20328%20772%20321%20761%20328Z%22%20fill%3D%22%23ffffff%22%2F%3E%0A%3C%2Fsvg%3E">
@@ -198,7 +198,7 @@
 </style>
 </head>
 <body>
-<!--SHARE_DATA_SLOT--><!--DEMO_DATA_SLOT--><!-- 示例查看器:CI 在此注入 <script type="application/json" id="demo-data">{payload}</script>;生产分享与托管查看器保持此占位不替换。 --><!-- 自包含分享文件:share.rs 在此注入 <script id="share-data">{blob}</script>;托管查看器保持此占位注释不替换(→ 拖文件模式)。 -->
+<!--SHARE_DATA_SLOT--><!-- 自包含分享文件:share.rs 在此注入 <script id="share-data">{blob}</script>;托管查看器保持此占位注释不替换(→ 拖文件模式)。 -->
 <div id="gate" class="gate">
   <div class="gate-card">
     <div class="brand-row" style="margin-bottom:6px">
@@ -1230,25 +1230,27 @@ let selfContained = false;
   } catch (e) { /* 非法数据节点:回退拖文件模式 */ }
 })();
 document.getElementById("go").addEventListener("click", submit);
+
+// 口令可随 URL fragment 送达:`#<口令>`。`#` 之后的内容按 HTTP 规范**不会**发给
+// 服务器,因此链接/二维码分享时,密钥始终只存在于收件人浏览器本地。到达即自动解密,
+// 免去手输;随后清掉 hash,避免口令留在地址栏与浏览历史里。
+(function passphraseFromHash() {
+  const h = decodeURIComponent(location.hash.replace(/^#/, "")).trim();
+  if (!h || !selfContained) return;
+  // 解析期调用 replaceState 会被随后的片段导航覆盖(实测),故解密完成后再清,
+  // 并在 load 时兜底一次 —— 口令绝不该留在地址栏和浏览历史里。
+  const clearHash = () => {
+    if (location.hash) history.replaceState(null, "", location.pathname + location.search);
+  };
+  addEventListener("load", clearHash);
+  // 解析期与 load 时清都可能被浏览器随后的片段导航覆盖,故再补一次延迟清除。
+  setTimeout(clearHash, 60);
+  pw.value = h;
+  Promise.resolve(submit()).finally(clearHash);
+})();
 pw.addEventListener("keydown", e => { if (e.key === "Enter") submit(); });
 
 // ── QA / 评审用示例(#37)──────────────────────────────────────────────
-// 示例查看器:部署时把 demo-payload.json 注入成 <script type="application/json" id="demo-data">
-// —— 与真实分享注入 share-data 同一机制(非执行数据节点,不需要 CSP 哈希)。
-// 生产分享文件不含此节点,因此不背任何演示数据;查看器代码则完全一致。
-var demoNode = document.getElementById("demo-data");
-if (demoNode) {
-  try {
-    render(JSON.parse(demoNode.textContent));
-    // 演示数据是虚构的,必须显著标注 —— 否则这一页读起来就是一份真实病历。
-    var demoBar = document.createElement("div");
-    demoBar.setAttribute("role", "note");
-    demoBar.style.cssText = "position:sticky;top:0;z-index:50;background:#fef3c7;border-bottom:1px solid #fcd34d;" +
-      "color:#92400e;font-size:13px;font-weight:600;padding:9px 16px;text-align:center;line-height:1.5";
-    demoBar.textContent = "示例演示 · 患者「张建国」及其全部化验、用药、影像、病理均为虚构,非真实病历。功能与真实分享完全一致,仅数据不同。";
-    document.body.insertBefore(demoBar, document.body.firstChild);
-  } catch (e) { console.error(e); }
-}
 </script>
 </body>
 </html>
